refactor session viewer actions
