+++ app/src/views/features/sessions/SessionViewer/DraftSessionViewer.tsx
@@ -1,4 +1,4 @@
-import React, { useCallback, useEffect, useMemo, useState } from "react";
+import React, { useEffect, useMemo, useState } from "react";
 import { useDispatch, useSelector } from "react-redux";
 import { unstable_usePrompt, useNavigate, useParams, useLocation } from "react-router-dom";
 import { getIsMiscTourCompleted, getUserAttributes, getUserAuthDetails } from "store/selectors";
@@ -9,10 +9,11 @@ import { ExclamationCircleOutlined } from "@ant-design/icons";
 import { RQButton } from "lib/design-system/components";
 import SessionDetails from "./SessionDetails";
 import { SessionViewerTitle } from "./SessionViewerTitle";
-import { RQSession, RQSessionAttributes, RQSessionEvents } from "@requestly/web-sdk";
+import { RQSession } from "@requestly/web-sdk";
 import mockSession from "./mockData/mockSession";
 import { ReactComponent as DownArrow } from "assets/icons/down-arrow.svg";
-import { compressEvents, decompressEvents, filterOutLargeNetworkResponses } from "./sessionEventsUtils";
+import { decompressEvents, filterOutLargeNetworkResponses, generateDraftSessionTitle } from "./sessionEventsUtils";
+import { cacheDraftSession, clearDraftSessionCache } from "./SessionViewerActions";
 import PageLoader from "components/misc/PageLoader";
 import { getSessionRecordingMetaData } from "store/features/session-recording/selectors";
 import { sessionRecordingActions } from "store/features/session-recording/slice";
@@ -24,14 +25,11 @@ import APP_CONSTANTS from "config/constants";
 import { ProductWalkthrough } from "components/misc/ProductWalkthrough";
 import { MISC_TOURS, TOUR_TYPES } from "components/misc/ProductWalkthrough/constants";
 import {
-  trackDraftSessionAutoSaved,
   trackDraftSessionDiscarded,
   trackDraftSessionViewed,
   trackSessionRecordingFailed,
 } from "modules/analytics/events/features/sessionRecording";
 import "./sessionViewer.scss";
-
-const MAX_CACHED_DRAFT_SESSIONS = 3;
 
 const DraftSessionViewer: React.FC = () => {
   const { tabId } = useParams();
@@ -49,14 +47,6 @@ const DraftSessionViewer: React.FC = () => {
   const [isSavePopupVisible, setIsSavePopupVisible] = useState(false);
   const [isSaveSessionClicked, setIsSaveSessionClicked] = useState(false);
 
-  const generateDraftSessionTitle = useCallback((url: string) => {
-    const hostname = new URL(url).hostname.split(".").slice(0, -1).join(".");
-    const date = new Date();
-    const month = date.toLocaleString("default", { month: "short" });
-    const formattedDate = `${date.getDate()}${month}${date.getFullYear()}`;
-    return `${hostname}@${formattedDate}`;
-  }, []);
-
   const hasUserCreatedSessions = useMemo(
     () =>
       userAttributes?.num_sessions > 0 ||
@@ -68,53 +58,6 @@ const DraftSessionViewer: React.FC = () => {
       userAttributes?.num_sessions_saved_offline,
     ]
   );
-
-  const handleCacheDraftSession = useCallback(
-    async (attributes: RQSessionAttributes, events: RQSessionEvents) => {
-      if (!attributes && !events) return;
-
-      const sessionToCache = {
-        metadata: { name: generateDraftSessionTitle(attributes.url), ...attributes },
-        events: compressEvents(events),
-      };
-
-      const draftSessions = await StorageService().getRecord(APP_CONSTANTS.DRAFT_SESSIONS);
-      if (!draftSessions) {
-        StorageService()
-          .saveRecord({ [APP_CONSTANTS.DRAFT_SESSIONS]: { [tabId]: sessionToCache } })
-          .then(() => {
-            trackDraftSessionAutoSaved();
-          })
-          .catch((e) => console.log(e));
-      } else {
-        if (tabId in draftSessions) {
-          delete draftSessions[tabId];
-        }
-
-        if (Object.keys(draftSessions).length === MAX_CACHED_DRAFT_SESSIONS) {
-          const oldestDraftKey = Object.keys(draftSessions)[0];
-          delete draftSessions[oldestDraftKey];
-        }
-        StorageService()
-          .saveRecord({
-            [APP_CONSTANTS.DRAFT_SESSIONS]: { ...draftSessions, [tabId]: sessionToCache },
-          })
-          .then(() => {
-            trackDraftSessionAutoSaved();
-          })
-          .catch((e) => console.log(e));
-      }
-    },
-    [tabId, generateDraftSessionTitle]
-  );
-
-  const handleClearDraftSessionCache = useCallback(async () => {
-    const draftSessions = await StorageService().getRecord(APP_CONSTANTS.DRAFT_SESSIONS);
-    if (draftSessions && tabId in draftSessions) {
-      delete draftSessions[tabId];
-      StorageService().saveRecord({ [APP_CONSTANTS.DRAFT_SESSIONS]: { ...draftSessions } });
-    }
-  }, [tabId]);
 
   useEffect(() => {
     const unloadListener = (e: BeforeUnloadEvent) => {
@@ -166,7 +109,7 @@ const DraftSessionViewer: React.FC = () => {
 
       dispatch(sessionRecordingActions.setEvents(mockSession.events));
       setIsLoading(false);
-    } else if (queryParams.has("saved")) {
+    } else if (queryParams.has("savedDraft")) {
       StorageService()
         .getRecord(APP_CONSTANTS.DRAFT_SESSIONS)
         .then((sessions) => {
@@ -201,13 +144,13 @@ const DraftSessionViewer: React.FC = () => {
 
             filterOutLargeNetworkResponses(tabSession.events);
             dispatch(sessionRecordingActions.setEvents(tabSession.events));
-            handleCacheDraftSession(tabSession.attributes, tabSession.events);
+            cacheDraftSession(tabSession.attributes, tabSession.events, tabId);
           }
         }
         setIsLoading(false);
       });
     }
-  }, [dispatch, tabId, user?.details?.profile?.email, generateDraftSessionTitle, queryParams, handleCacheDraftSession]);
+  }, [dispatch, tabId, user?.details?.profile?.email, queryParams]);
 
   const confirmDiscard = () => {
     Modal.confirm({
@@ -217,7 +160,7 @@ const DraftSessionViewer: React.FC = () => {
       okText: "Yes",
       cancelText: "No",
       onOk: async () => {
-        await handleClearDraftSessionCache();
+        await clearDraftSessionCache(tabId);
         trackDraftSessionDiscarded();
         navigate(PATHS.SESSIONS.ABSOLUTE);
       },
@@ -260,7 +203,6 @@ const DraftSessionViewer: React.FC = () => {
             <SaveRecordingConfigPopup
               onClose={() => setIsSavePopupVisible(false)}
               setIsSaveSessionClicked={setIsSaveSessionClicked}
-              clearSavedDraft={handleClearDraftSessionCache}
             />
           )}
         </div>
