+++ app/src/views/features/sessions/SessionViewer/constants.ts
@@ -1,2 +1,3 @@
 export const SESSION_EXPORT_TYPE = "SESSION";
 export const EXPORTED_SESSION_FILE_EXTENSION = "rqly";
+export const MAX_CACHED_DRAFT_SESSIONS = 3;
