+++ app/src/views/features/sessions/SessionViewer/sessionEventsUtils.ts
@@ -125,3 +125,11 @@ export const downloadSession = (fileContent: string, fileName: string): void => 
 export const getSessionRecordingOptions = (options: RecordingOptions): string[] => {
   return Object.keys(options ?? {}).filter((key: DebugInfo) => options?.[key]);
 };
+
+export const generateDraftSessionTitle = (url: string) => {
+  const hostname = new URL(url).hostname.split(".").slice(0, -1).join(".");
+  const date = new Date();
+  const month = date.toLocaleString("default", { month: "short" });
+  const formattedDate = `${date.getDate()}${month}${date.getFullYear()}`;
+  return `${hostname}@${formattedDate}`;
+};
