+++ app/src/views/features/sessions/SessionsIndexPageContainer/SessionsIndexPage/index.js
@@ -75,6 +75,7 @@ const SessionsIndexPage = () => {
           visibility: Visibility.ONLY_ME,
         };
       });
+      console.log({ formattedRecords });
       return formattedRecords;
     }
